# api/openapi.yaml
openapi: 3.0.0
info:
  title: CBT API
  version: 1.0.0
  description: ClickHouse Build Tool API for managing transformation models
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /models:
    get:
      summary: List all models (lightweight)
      operationId: listAllModels
      tags:
        - Models
      description: |
        Returns a lightweight list of all models across all types.
        Use this for navigation, search, or building dependency graphs.
        For full model details, use type-specific endpoints.
      parameters:
        - name: type
          in: query
          description: Filter by model type
          schema:
            type: string
            enum: [external, transformation]
        - name: database
          in: query
          description: Filter by database name
          schema:
            type: string
          example: analytics
        - name: search
          in: query
          description: Search by model ID or description (case-insensitive)
          schema:
            type: string
          example: daily
      responses:
        '200':
          description: List of all models (summary view)
          content:
            application/json:
              schema:
                type: object
                required: [models, total]
                properties:
                  models:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelSummary'
                  total:
                    type: integer
              example:
                models:
                  - id: sources.events
                    type: external
                    database: sources
                    table: events
                    description: Raw event data from production
                  - id: analytics.daily_users
                    type: transformation
                    database: analytics
                    table: daily_users
                    description: Daily active user counts
                total: 2
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # EXTERNAL MODELS
  # ============================================
  /models/external:
    get:
      summary: List external models
      operationId: listExternalModels
      tags:
        - External Models
      parameters:
        - name: database
          in: query
          description: Filter by database name
          schema:
            type: string
          example: sources
      responses:
        '200':
          description: List of external models
          content:
            application/json:
              schema:
                type: object
                required: [models, total]
                properties:
                  models:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExternalModel'
                  total:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
  /models/external/{id}:
    get:
      summary: Get external model by ID
      operationId: getExternalModel
      tags:
        - External Models
      parameters:
        - name: id
          in: path
          required: true
          description: Fully qualified model ID (database.table)
          schema:
            type: string
          example: sources.clickhouse_prod
      responses:
        '200':
          description: External model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalModel'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # TRANSFORMATION MODELS
  # ============================================
  /models/transformations:
    get:
      summary: List transformation models
      operationId: listTransformations
      tags:
        - Transformations
      parameters:
        - name: database
          in: query
          description: Filter by database name
          schema:
            type: string
          example: analytics
        - name: type
          in: query
          description: Filter by transformation type
          schema:
            type: string
            enum: [scheduled, incremental]
        - name: status
          in: query
          description: Filter by last run status
          schema:
            type: string
            enum: [success, failed, running, pending]
      responses:
        '200':
          description: List of transformation models
          content:
            application/json:
              schema:
                type: object
                required: [models, total]
                properties:
                  models:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransformationModel'
                  total:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
  /models/transformations/{id}:
    get:
      summary: Get transformation model by ID
      operationId: getTransformation
      tags:
        - Transformations
      parameters:
        - name: id
          in: path
          required: true
          description: Fully qualified model ID (database.table)
          schema:
            type: string
          example: analytics.daily_active_users
      responses:
        '200':
          description: Transformation model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransformationModel'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # EXTERNAL MODEL BOUNDS
  # ============================================
  /models/external/bounds:
    get:
      summary: List external model bounds
      operationId: listExternalBounds
      tags:
        - External Models
      description: Returns min/max positions from Redis cache for all external models
      responses:
        '200':
          description: List of external model bounds
          content:
            application/json:
              schema:
                type: object
                required: [bounds, total]
                properties:
                  bounds:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExternalBounds'
                  total:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
  /models/external/{id}/bounds:
    get:
      summary: Get external model bounds by ID
      operationId: getExternalBounds
      tags:
        - External Models
      description: Returns min/max positions and scan metadata from Redis cache
      parameters:
        - name: id
          in: path
          required: true
          description: Fully qualified model ID (database.table)
          schema:
            type: string
          example: ethereum.beacon_blocks
      responses:
        '200':
          description: External model bounds details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalBounds'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # TRANSFORMATION COVERAGE
  # ============================================
  /models/transformations/coverage:
    get:
      summary: List transformation coverage
      operationId: listTransformationCoverage
      tags:
        - Transformations
      description: Returns processed ranges from admin_incremental table for all incremental transformations
      parameters:
        - name: database
          in: query
          description: Filter by database name
          schema:
            type: string
          example: analytics
      responses:
        '200':
          description: List of transformation coverage
          content:
            application/json:
              schema:
                type: object
                required: [coverage, total]
                properties:
                  coverage:
                    type: array
                    items:
                      $ref: '#/components/schemas/CoverageSummary'
                  total:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
  /models/transformations/{id}/coverage:
    get:
      summary: Get transformation coverage by ID
      operationId: getTransformationCoverage
      tags:
        - Transformations
      description: Returns all processed ranges from admin_incremental table
      parameters:
        - name: id
          in: path
          required: true
          description: Fully qualified model ID (database.table)
          schema:
            type: string
          example: analytics.daily_users
      responses:
        '200':
          description: Transformation coverage details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageDetail'
        '400':
          description: Model is not incremental type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # SCHEDULED RUNS
  # ============================================
  /models/transformations/runs:
    get:
      summary: List scheduled transformation runs
      operationId: listScheduledRuns
      tags:
        - Transformations
      description: Returns last run timestamps from admin_scheduled table for all scheduled transformations
      parameters:
        - name: database
          in: query
          description: Filter by database name
          schema:
            type: string
          example: analytics
      responses:
        '200':
          description: List of scheduled transformation runs
          content:
            application/json:
              schema:
                type: object
                required: [runs, total]
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduledRun'
                  total:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
  /models/transformations/{id}/runs:
    get:
      summary: Get scheduled transformation run by ID
      operationId: getScheduledRun
      tags:
        - Transformations
      description: Returns last run timestamp from admin_scheduled table
      parameters:
        - name: id
          in: path
          required: true
          description: Fully qualified model ID (database.table)
          schema:
            type: string
          example: analytics.daily_report
      responses:
        '200':
          description: Scheduled transformation run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledRun'
        '400':
          description: Model is not scheduled type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
...
    ModelSummary:
      type: object
      required: [id, type, database, table]
      description: Lightweight model representation for listings
      properties:
        id:
          type: string
          description: Fully qualified ID (database.table)
          example: analytics.daily_users
        type:
          type: string
          enum: [external, transformation]
          description: Model type
        database:
          type: string
          example: analytics
        table:
          type: string
          example: daily_users
        description:
          type: string
          example: Daily active user counts

    # ============================================
    # EXTERNAL MODEL SCHEMAS
    # ============================================
    ExternalModel:
      type: object
      required: [id, database, table]
      properties:
        id:
          type: string
          description: Fully qualified ID (database.table)
          example: ethereum.beacon_blocks
        database:
          type: string
          example: ethereum
        table:
          type: string
          example: beacon_blocks
        description:
          type: string
          example: Beacon chain blocks from production cluster
        cache:
          type: object
          description: Cache configuration for external source
          properties:
            incremental_scan_interval:
              type: string
              description: Interval for incremental cache updates
              example: "10s"
            full_scan_interval:
              type: string
              description: Interval for full cache refresh
              example: "1h"
        lag:
          type: integer
          description: Number of blocks/slots to lag behind head
          example: 10
        metadata:
          type: object
          readOnly: true
          description: System-managed metadata
          properties:
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            last_synced_at:
              type: string
              format: date-time
            row_count:
              type: integer
              description: Approximate row count in destination table
            size_bytes:
              type: integer
              description: Approximate table size in bytes

    # ============================================
    # TRANSFORMATION MODEL SCHEMAS
    # ============================================

    # Base schema with common fields
...
    TransformationModel:
      allOf:
        - $ref: '#/components/schemas/TransformationModelBase'
        - type: object
          properties:
            # Scheduled-specific fields (optional)
            schedule:
              type: string
              description: Cron expression (present when type=scheduled)
              example: "@every 5m"
            # Incremental-specific fields (optional)
            interval:
              type: object
              description: Interval configuration (present when type=incremental)
              properties:
                min:
                  type: integer
                  description: Minimum interval size
                max:
                  type: integer
                  description: Maximum interval size
            schedules:
              type: object
              description: Schedules (present when type=incremental)
              properties:
                forwardfill:
                  type: string
                  description: Forward fill schedule
                backfill:
                  type: string
                  description: Backfill schedule
            limits:
              type: object
              description: Limits (present when type=incremental)
              properties:
                min:
                  type: integer
                  description: Minimum position limit
                max:
                  type: integer
                  description: Maximum position limit

    # ============================================
    # BOUNDS, COVERAGE & RUNS SCHEMAS
    # ============================================

    # External model bounds from Redis cache
    ExternalBounds:
      type: object
      required: [id, min, max]
      properties:
        id:
          type: string
          description: Fully qualified model ID (database.table)
          example: ethereum.beacon_blocks
        min:
          type: integer
          description: Minimum position in external source
          example: 0
        max:
          type: integer
          description: Maximum position in external source
          example: 10445672
        previous_min:
          type: integer
          description: Previous minimum position (for tracking changes)
          example: 0
        previous_max:
          type: integer
          description: Previous maximum position (for tracking changes)
          example: 10445662
        last_incremental_scan:
          type: string
          format: date-time
          description: Timestamp of last incremental scan
        last_full_scan:
          type: string
          format: date-time
          description: Timestamp of last full scan
        initial_scan_complete:
          type: boolean
          description: Whether initial scan is complete
        initial_scan_started:
          type: string
          format: date-time
          description: Timestamp when initial scan started

    # Position range from admin table
...
    CoverageSummary:
      type: object
      required: [id, ranges]
      properties:
        id:
          type: string
          description: Fully qualified model ID (database.table)
          example: analytics.daily_users
        ranges:
          type: array
          items:
            $ref: '#/components/schemas/Range'
          description: Processed ranges from admin_incremental table

    # Transformation coverage detail (for detail endpoint)
    CoverageDetail:
      type: object
      required: [id, ranges]
      properties:
        id:
          type: string
          description: Fully qualified model ID (database.table)
          example: analytics.daily_users
        ranges:
          type: array
          items:
            $ref: '#/components/schemas/Range'
          description: All processed ranges from admin_incremental table

    # Scheduled transformation run info
    ScheduledRun:
      type: object
      required: [id]
      properties:
        id:
          type: string
          description: Fully qualified model ID (database.table)
          example: analytics.daily_report
        last_run:
          type: string
          format: date-time
          description: Timestamp of last execution

    # ============================================
    # ERROR SCHEMAS
    # ============================================
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: integer
          description: HTTP status code
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Model not found: analytics.nonexistent"
            code: 404
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: 500